+++ honglidong/app/src/main/java/com/hongildong/map/ui/search/SearchedItem.kt
@@ -24,10 +24,11 @@ import com.hongildong.map.ui.theme.Gray200
 import com.hongildong.map.ui.theme.HongildongTheme
 
 @Composable
-fun RecentlySearchedItem(
+fun SearchedItem(
     itemName: String,
     onClickItem: (String) -> Unit,
-    onDeleteItem: (String) -> Unit
+    onDeleteItem: (String) -> Unit,
+    isRecentlySearched: Boolean = false
 ) {
     Column {
         Row (
@@ -53,16 +54,18 @@ fun RecentlySearchedItem(
                     style = AppTypography.Regular_15
                 )
             }
-            // 삭제 버튼
-            Image(
-                painter = painterResource(R.drawable.ic_close),
-                contentDescription = stringResource(R.string.delete_item, itemName),
-                contentScale = ContentScale.Fit,
-                modifier = Modifier
-                    .padding(4.dp)
-                    .size(12.dp)
-                    .clickable { onDeleteItem(itemName) }
-            )
+            if (isRecentlySearched) {
+                // 삭제 버튼
+                Image(
+                    painter = painterResource(R.drawable.ic_close),
+                    contentDescription = stringResource(R.string.delete_item, itemName),
+                    contentScale = ContentScale.Fit,
+                    modifier = Modifier
+                        .padding(4.dp)
+                        .size(12.dp)
+                        .clickable { onDeleteItem(itemName) }
+                )
+            }
         }
 
         HorizontalDivider(
@@ -77,20 +80,16 @@ fun RecentlySearchedItem(
 fun RecentlySearchedItemPreview() {
     HongildongTheme {
         Column {
-            RecentlySearchedItem(
+            SearchedItem(
                 itemName = "아이템1",
                 onClickItem = {},
                 onDeleteItem = {}
             )
-            RecentlySearchedItem(
+            SearchedItem(
                 itemName = "아이템2",
                 onClickItem = {},
-                onDeleteItem = {}
-            )
-            RecentlySearchedItem(
-                itemName = "아이템3",
-                onClickItem = {},
-                onDeleteItem = {}
+                onDeleteItem = {},
+                isRecentlySearched = true
             )
         }
     }
